Add files via upload
Data is always embedded in a context, and it is necessary to understand that context and know where the data comes from and whom it represents. The dataset(https://data.peclet.com.au/explore/dataset/nsw-domestic-waste-generation-and-recycling-rates-from-2015/export/?disjunctive.council_name) "NSW domestic waste generation and recycling rates - from 2015" used in this analysis was found in one of the open databases of Peclet Technology (Peclet), a firm founded by technology consultants and data scientists with the vision to develop and bring innovative IoT and data-sharing solutions to Australia. One of their main focuses is smart cities, and most of their projects include solutions that combine real-time data directly from IoT sensors. However, this dataset is part of the data used to create reports about domestic waste and recycling in NSW. The NSW EPA (NSW Environment Protection Authority) provides different [reports](https://www.epa.nsw.gov.au/your-environment/recycling-and-reuse/warr-strategy/policy-makers/surveys) on waste and recyclables collected from NSW households since 2006 and until 2021. According to Peclet, part of this data was processed and the new dataset was created, containing some data on waste generation and recycling rates - from 2015 to 2021.

Peclet claims to have added a breakdown by material type of the recycled and residual waste volumes as an indicative breakdown. When going deeper into the NWS EPA website, they claim that when comparing the figures for local government areas (LGAs), it is important to consider regional variations in consumption patterns, available services, data availability, and the interpretation of survey questions.

Some limitations of the data include missing data in various columns, including the `geolocation` column, making it impossible to visualize some councils in the plots and maps. In addition, since Peclet is not the primary data source, we had to look deeper to see the meaning and purpose of some variables whose description was unavailable on Peclet's website. However, this data have gaps. For instance,  Peclet does not mention what pre-processing they did while collecting the data from NSW EPA. Furthermore, it is also not easy to identify which criteria the NSW EPA uses to split the councils within regions.

Finally, the data available in this data set and even going deeper into the NSW EPA website, which is the primary source, has only data until the financial year 2021, which make our analysis less relevant given the actual context.